drm/imagination: Rework use of pvr_job structure
The pvr_job structure was not really being used; it contained very
few relevant members and only have a lifetime of the job submit
command. This commit reworks the structure to contain the job command
data, ID, pointers to the context and device, be tracked within the
device, and to persist until job completion (signalled by the output
fence).

Signed-off-by: Sarah Walker <sarah.walker@imgtec.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>