drm/imagination: Fix use-after-free in the VM unmap path
When the last entry in an L0/L1 page table is killed, we destroy the
page table, but we can't really free the page table until we're sure

1. the page-table walk cache doesn't have any ref to the L1/L0 page
   tables we're destroying, otherwise the GPU might trigger a read
   access to a page that has been returned to the system and possibly
   re-allocated to someone else. This requires flushing the PD/PT
   caches before we destroy the page table.
2. we cleared any reference to the L0/L1 page tables the
   pvr_page_table_ptr object might have, otherwise we end up with a
   use-after-free situation.

Item 2 is easily solved by resetting the ->l{0,1}_table pointer in
pvr_page_table_l{0,1}_remove(), and only calling
pvr_page_table_ptr_sync_manual() if the l1/l0 table is != NULL.
Note that we don't need to flush the page table if it's being destroyed,
so skipping the synchronization in that case should be fine.

Item 1 could be solved by flushing the PT/PD every time a page table is
removed and before it's destroyed. But we probably want to issue a single
flush when the unmap operation covers more than one PT. So let's
collect all L0/L1 tables in the ptr object, and free those in
pvr_page_table_ptr_fini(), after we've flushed the PC, PD and TLB
caches. With pvr_vm_mmu_flush() being called in pvr_vm_mmu_flush(),
we no longer need to call it in pvr_vm_mapping_[un]map().

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>