drm/imagination: Fix the fence consolidation in the null_job handling
We want the fence returned by pvr_process_job_null() to be assigned a
unique context ID, otherwise fence_array_add() will consider that
fences attached to this context are signaled in-order, which is not
guaranteed here. So let's allocate a new fence context for each
fence_array we create.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>