drm/imagination: Add infra to support passing arrays of complex objects
It is not unusual to have to extend objects to pass new parameters, and
the DRM ioctl infrastructure is supporting that by padding ioctl
arguments with zeros when the data passed by userspace is smaller than
the struct defined in the drm_ioctl_desc, thus keeping things backward
compatible. This drm_pvr_obj_array is just applying the same concepts
to indirect objects passed through arrays referenced from the main ioctl
arguments structure: the stride basically defines the size of the object
passed by userspace, which allows the kernel driver to pad things with
zeros when it's smaller than the size of the object it expects.

Userspace should use DRM_PVR_OBJ_ARRAY() to fill object array fields, so
it doesn't have to manually set the stride. On the kernel side, we
provide a macro to easily define array getters based on the object name
and the last mandatory field in this object. When adding a new structure,
the last field should always be the actual last field in the struct, and
should never change after that.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>