drm/imagination: Add a drm_pvr_sync_op object to the uAPI
Add a drm_pvr_sync_op object to the uAPI so we can pass more than just
handles, which is needed if we want to support timeline sync objects,
among other things. We also take this as an opportunity to merge the
out/in sync objects into a single array, and provide the action to do
on the syncobj (signal or wait) as a flag.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>