drm/imagination: Fix the context teardown logic
Right now, contexts are first waited on for a bit, then all their
jobs are being cancelled and finally, the context is destroyed.

This lead to potential races between the job submission and the job
cancellation part, and it also means we're potentially teardown FW
context objects before the FW is done with them.

The first sort of race is addressed by letting
pvr_context_queue_check_pending_jobs() deal with pending job
cancellation, all pvr_context_destroy() has to do is flag the context
as destroyed and queue a pending job event. We also check the destroyed
flag to make sure no new jobs are submitted to the hardware. If the
submit job path acquired the context before it was removed from the
ctx_handles set, we'll just let it queue the jobs and trigger the
job event, which will result in pvr_context_queue_check_pending_jobs()
immediately cancelling them.

Second issue is addressed by keeping in_flight jobs around until the
firmware informs us it's done with them. Since each job owns a reference
to the context, the context will stay around until all in-flight jobs
are signaled. This is far from ideal, because we don't have any watchdog
to make sure we don't leak the context. But let's keep this robustness
improvement for later.

We do signal done_fence objects attached to in-flight jobs as soon as
the context is destroyed, to unblock any other objects waiting on them.
We take care of setting the fence error to ECANCELED in that case.

So that's it. Hopefully it's good enough to get rid of those backtraces
when a program crashes and jobs are still in flight. And things will soon
be improved with extra robustness mechanisms (timeouts triggering a
GPU+FW reset and the eviction of all blocked contexts).

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>