drm/sched: Add native dependency support to drm_sched
A native dependency is a dma_fence object the GPU driver can natively
wait on. Often, it is just a firmware construct that allowing the host
CPU to delegate synchronization to the micro-controller in charge of
job scheduling. For this to work properly, we need drm_sched to be
aware of these native dependencies so it can CPU-wait on those when
jobs are killed, but ignore those dependencies in the dequeuing path.

Signed-off-by: Donald Robson <donald.robson@imgtec.com>