drm/imagination: Make sure we have a KCCB slot available to queue the CMD_KICK
We need to do our best to guarantee that pvr_queue_run_job() doesn't
fail. We already describe the lack of CCCB space as an internal
dependency, returning a dma_fence that's signaled when enough space is
available to queue our command. Let's do the same for the KCCB slot
that's needed to issue our CMD_KICK.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>