drm/imagination: Make queue fences retain a ref on the context
In theory, dma_fence objects returned to drm_sched through
{run,prepare}_job() are not leaving the
drm_sched_entity/drm_gpu_scheduler scope, and should be signaled when
the entity/scheduler they come from is killed. But, with native fences
it gets trickier, because jobs on one queue might reference the UFO
object on another queue and we want this UFO object to stay pinned as
long as we have native fences pointing to it.

This is easily solved by grabbing a reference to the context when
initializing a queue fence, and releasing it when the fence is
destroyed.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>