drm/sched: Kill the loop in drm_sched_main()
This way we introduce a tiny bit of fairness between schedulers assigned
to the same workqueue. Without it, we might have one scheduler
monopolizing the workqueue leaving others no chance to submit their
jobs.

Note: we might want to keep the loop with a limit on the number of
iterations.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>