drm/imagination: Fix unbalanced dma_fence_{get,put} in pvr_job_add_deps()
drm_sched_job_add_dependency() doesn't grab a reference on the fence
it's being passed when adding it to the dependency array. It's expecting
the caller to do that, and it's also releasing the ref for us if the
fence is not kept. So, all we have to do here is grab a reference to
the unwrapped_fence fence before passing it to
drm_sched_job_add_dependency().

What's missing though, is a dma_fence_put() on the main fence, because
we didn't pass this one to drm_sched_job_add_dependency() and we have
to release it explicitly.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>