drm/imagination: Fix a race between signal_done_fences() and timedout_job()
Before we stop the scheduler, make sure the queue is out of any list, so
any call to pvr_queue_update_active_state_locked() that might happen
until the scheduler is really stopped doesn't end up re-inserting the
queue in the active list. This would cause
pvr_queue_signal_done_fences() and drm_sched_stop() to race with each
other when accessing the pending_list, since drm_sched_stop() doesn't
grab the job_list_lock when modifying the list (it's assuming the
only other accessor is the scheduler, and it's safe to not grab the
lock since it's stopped).

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>