drm/imagination: Move to a threaded IRQ
There are several benefits to moving to a threaded IRQ model:

* we can coalesce interrupts and avoid interrupting the system too often
  when the GPU gets passed a lot of short-lived jobs that takes less
  time to execute than it takes time for the host CPU to process the
  FW events
* if we move the heavy processing to a thread, we no longer need the
  irq_wq and can instead call the FWCCB/KCCB/queue processing functions
  directly
* we are a bit more fair to the rest of the system

There's of course an extra latency induced, but given everything except
the rtn_q wake_up was delegated to a workqueue, we shouldn't notice
a big difference. And if we really care about speeding up the
rtn_q wake_up(), we can always move this call back to the hard IRQ
handler.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>