dt-bindings: gpu: img: remove excess address and size cells in example
The number of register address and size cells both default to 1, but 2 of each
were being used in the example for ti,am62-gpu. This resulted in the following
output when running `make dt_binding_check DT_SCHEMA_FILES=img,powervr.yaml`:

    ./Documentation/devicetree/bindings/gpu/img,powervr.example.dtb: gpu@fd00000: reg: [[0, 265289728], [0, 131072]] is too long
            from schema $id: http://devicetree.org/schemas/gpu/img,powervr.yaml#

As the address bus width for the SoC interface is 32 bits, fix this by removing
the use of the additional cells.

Signed-off-by: Frank Binns <frank.binns@imgtec.com>