drm/imagination: Fix context release in final error path in pvr_context_create()
The final error path in pvr_context_create() occurs after the context has
been added to the device context xarray. This means it is theoretically
possible that another thread could look up the context and take a reference
before the function returns. We need to account for this possibility. In this
error path, instead of directly destroying the context, drop our reference on
the context instead.

Signed-off-by: Sarah Walker <sarah.walker@imgtec.com>